src/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile requirements.in
#
apimapper==0.7.8
    # via -r requirements.in
apispec==0.39.0
    # via -r requirements.in
asgiref==3.7.2
    # via django
async-timeout==4.0.3
    # via redis
attrs==23.2.0
    # via jsonschema
bibtexparser==1.2.0
    # via -r requirements.in
build==1.0.3
    # via pip-tools
certifi==2023.11.17
    # via
    #   requests
    #   sentry-sdk
charset-normalizer==3.3.2
    # via requests
clamd==1.0.2
    # via -r requirements.in
click==8.1.7
    # via
    #   pip-tools
    #   rq
concurrent-log-handler==0.9.19
    # via -r requirements.in
crontab==1.0.1
    # via rq-scheduler
django==3.2.23
    # via
    #   -r requirements.in
    #   django-cas-ng
    #   django-cors-headers
    #   django-debug-toolbar
    #   django-extensions
    #   django-filter
    #   django-redis
    #   django-rq
    #   djangorestframework
    #   drf-yasg
django-cas-ng==4.3.0
    # via -r requirements.in
django-cleanup==5.2.0
    # via -r requirements.in
django-cors-headers==3.7.0
    # via -r requirements.in
django-debug-toolbar==3.2.1
    # via -r requirements.in
django-environ==0.4.5
    # via -r requirements.in
django-extensions==3.2.3
    # via -r requirements.in
django-filter==23.2
    # via -r requirements.in
django-redis==4.12.1
    # via -r requirements.in
django-rq==2.8.1
    # via -r requirements.in
djangorestframework==3.13.1
    # via
    #   -r requirements.in
    #   drf-yasg
drf-yasg[validation]==1.21.7
    # via -r requirements.in
freezegun==1.4.0
    # via rq-scheduler
future==0.18.3
    # via bibtexparser
gunicorn==20.1.0
    # via -r requirements.in
hashids==1.3.1
    # via -r requirements.in
idna==3.6
    # via
    #   jsonschema
    #   requests
importlib-metadata==7.0.1
    # via
    #   build
    #   markdown
inflection==0.5.1
    # via drf-yasg
isodate==0.6.1
    # via rdflib
jsonpointer==2.4
    # via jsonschema
jsonschema[format]==3.2.0
    # via
    #   -r requirements.in
    #   swagger-spec-validator
lxml==5.1.0
    # via python-cas
markdown==3.4.4
    # via
    #   -r requirements.in
    #   apimapper
marshmallow==2.21.0
    # via -r requirements.in
packaging==23.2
    # via
    #   build
    #   drf-yasg
pip-tools==7.3.0
    # via -r requirements.in
portalocker==2.8.2
    # via concurrent-log-handler
progressbar2==3.53.1
    # via -r requirements.in
psutil==5.9.7
    # via rainbow-saddle
psycopg2-binary==2.8.6
    # via -r requirements.in
pyparsing==3.1.1
    # via
    #   bibtexparser
    #   rdflib
pyproject-hooks==1.0.0
    # via build
pyrsistent==0.20.0
    # via jsonschema
python-cas==1.6.0
    # via django-cas-ng
python-dateutil==2.8.2
    # via
    #   freezegun
    #   rq-scheduler
python-magic==0.4.22
    # via -r requirements.in
python-utils==3.8.1
    # via progressbar2
pytz==2023.3.post1
    # via
    #   django
    #   djangorestframework
    #   drf-yasg
pyyaml==6.0.1
    # via
    #   apispec
    #   drf-yasg
    #   swagger-spec-validator
rainbow-saddle==0.4.0
    # via -r requirements.in
rdflib==5.0.0
    # via
    #   -r requirements.in
    #   skosmos-client
redis==5.0.1
    # via
    #   django-redis
    #   django-rq
    #   rq
requests==2.31.0
    # via
    #   apimapper
    #   python-cas
    #   skosmos-client
rfc3987==1.3.8
    # via jsonschema
rq==1.15.1
    # via
    #   django-rq
    #   rq-scheduler
rq-scheduler==0.13.1
    # via -r requirements.in
sentry-sdk==1.39.1
    # via -r requirements.in
shortuuid==1.0.8
    # via -r requirements.in
six==1.16.0
    # via
    #   isodate
    #   jsonschema
    #   progressbar2
    #   python-cas
    #   python-dateutil
    #   rdflib
skosmos-client==0.3.0
    # via -r requirements.in
sqlparse==0.4.4
    # via
    #   django
    #   django-debug-toolbar
strict-rfc3339==0.7
    # via jsonschema
swagger-spec-validator==3.0.3
    # via drf-yasg
titlecase==2.3
    # via -r requirements.in
tomli==2.0.1
    # via
    #   build
    #   pip-tools
    #   pyproject-hooks
typing-extensions==4.9.0
    # via
    #   asgiref
    #   python-utils
    #   swagger-spec-validator
uritemplate==4.1.1
    # via drf-yasg
urllib3==2.1.0
    # via
    #   requests
    #   sentry-sdk
webcolors==1.13
    # via jsonschema
wheel==0.42.0
    # via pip-tools
xmltodict==0.12.0
    # via -r requirements.in
zipp==3.17.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
